// src/app/modules/graficos/models/category.ts
export interface Category{
    customizacao?: number;
    treinamentoOnline?: number;
    implantacao?: number;
    licitacao?: number;
    solicitacaoServico?: number;
    homologacao?: number;
    falha?: number;
    semCategoria?: number;
    solicitacaoTreinamento?: number;
    sugestao?: number;
    duvida?: number;
}